fix: improve error handling for user-/meta-/vendor-data handlers (#93)
* fix(SMDClient): GroupMembership: returhn error if id is empty

Signed-off-by: Devon Bautista <17506592+synackd@users.noreply.github.com>

* feat(SMDClient): add ErrEmptyID and ErrSMDResponse errors

ErrEmptyID is returned if an ID (xname) passed is empty. Functions can
check for this error to warn the caller not to pass an empty ID.

ErrSMDResponse is an error that wraps an *http.Response, specifically
one that is returned from calls to SMD (e.g. via getSMD()). This is so
callers of getSMD() can distinguish between control flow errors and HTTP
errors (response >= 400) from SMD.

This commit also separates error definitions for the smdclient package
into their own errors.go file.

Signed-off-by: Devon Bautista <17506592+synackd@users.noreply.github.com>

* fix(metadata_handlers): distinguish HTTP errors from control flow errors

In MetaDataHandler(), issue HTTP response based on if the errors
returned by smd.ComponentInformation() and smd.GroupMembership() are
control flow errors or unsuccessful HTTP return codes (>= 400). With the
addition of ErrSMDResponse, the error can be checked if it is an HTTP
error from an API call to SMD or a control flow error and perform the
proper actions based on this.

Signed-off-by: Devon Bautista <17506592+synackd@users.noreply.github.com>

* fix(userdata_handlers): clarify not found error message

When asking for the user data of either:

- a non-existent group
- a non-existent node
- a node that is not in a group

the previous error message simply said "Group not found" which can be
misleading. This commit clarifies that the error could be any of the
above possibilities.

Signed-off-by: Devon Bautista <17506592+synackd@users.noreply.github.com>

* fix(vendordata_handlers): improve error messages

Signed-off-by: Devon Bautista <17506592+synackd@users.noreply.github.com>

---------

Signed-off-by: Devon Bautista <17506592+synackd@users.noreply.github.com>