feat: consistent base URL and fake SMD CLI
Add a --smd-simulator flag for the CLOUD_INIT_SMD_SIMULATOR and set the
fakeSMDEnabled variable to it. Previously the envvar was used in some
cases, and the variable in others, and nothing actually set the
variable.

Pass baseUrl to the VendorDataHandler initializer and use it if none is
set in the store. Previously this was only used for logging, and was
effectively inaccurate as such.

Signed-off-by: Travis Raines <571832+rainest@users.noreply.github.com>